+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- [updated] command support Chinese language.
+
 ## v0.0.2
 - Get plugin list from `this.app.plugins.plugins` rather than read `.obsidian/community-plugins.json` file.
 - [feature] generate table format